More work towards the generalized Constraint structure
- Expanded RodConstraint to set up J, JDot, C and CDot matrices and
vectors.
- First part of Equation 11 implemented in Solver.
- More const correctness in the RodConstraint
- Changed some floats to doubles and vice versa in Solver and TinkerToy
to fix some precision warnings.
- Changed some ints to longs in Solver and TinkerToy to fix some
precision warnings.
- Removed unnecessary break and added default case to remap_GUI switch.

+++ Solver.h
@@ -17,13 +17,13 @@ struct Derivative {
 
 enum class SolverType { Euler, Midpoint, RungeKutta4 };
 
-void simulation_step(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const float dt,const SolverType m_SolverType);
+void simulation_step(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const double dt, const SolverType m_SolverType);
 
-void EulerSolver(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const float dt);
+void EulerSolver(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const double dt);
 
-void MidpointSolver(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const float dt);
+void MidpointSolver(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const double dt);
 
-void RungeKutta4thOrderSolver(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const float dt);
+void RungeKutta4thOrderSolver(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, const double dt);
 
 void ParticleDerivative(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector, const std::vector<Force*> & cVector, std::vector<Derivative> & dVector);
 
@@ -31,4 +31,4 @@ void ClearForces(const std::vector<Particle*> & pVector);
 
 void CalculateForces(const std::vector<Particle*> & pVector, const std::vector<Force*> & fVector);
 
-void ScaleDerivativeVector(std::vector<Derivative> & dVector, const float scaleFactor);
+void ScaleDerivativeVector(std::vector<Derivative> & dVector, const double scaleFactor);
